+++ src/main/java/com/zerozero/global/s3/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.zerozero.global.config.s3;
+package com.zerozero.global.s3.config;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
